chore(netflow): dashmap dep, spellcheck words, and style alignment
- Optional dashmap dependency wired to sources-netflow feature
- Consolidate Protocol* internal events in events.rs; docs on configurable_component
- rustfmt and minor parser/formatting cleanup
