+++ Cargo.toml
@@ -140,6 +140,7 @@ chrono = { version = "0.4.41", default-features = false, features = ["clock", "s
 chrono-tz = { version = "0.10.3", default-features = false, features = ["serde"] }
 clap = { version = "4.5.40", default-features = false, features = ["derive", "error-context", "env", "help", "std", "string", "usage", "wrap_help"] }
 darling = { version = "0.20.11", default-features = false, features = ["suggestions"] }
+dashmap = { version = "6.1.0", default-features = false }
 flate2 = { version = "1.1.2", default-features = false, features = ["zlib-rs"] }
 futures = { version = "0.3.31", default-features = false, features = ["compat", "io-compat", "std"], package = "futures" }
 glob = { version = "0.3.2", default-features = false }
@@ -325,6 +326,7 @@ chrono.workspace = true
 chrono-tz.workspace = true
 colored = { version = "3.0.0", default-features = false }
 csv = { version = "1.3", default-features = false }
+dashmap = { workspace = true, optional = true }
 databend-client = { version = "0.27.1", default-features = false, features = ["rustls"], optional = true }
 derivative = { version = "2.2.0", default-features = false }
 dirs-next = { version = "2.0.0", default-features = false, optional = true }
@@ -643,7 +645,7 @@ sources-logstash = ["sources-utils-net-tcp", "tokio-util/net"]
 sources-mongodb_metrics = ["dep:mongodb"]
 sources-mqtt = ["dep:rumqttc"]
 sources-nats = ["dep:async-nats", "dep:nkeys"]
-sources-netflow = ["sources-utils-net-udp", "tokio-util/net", "dep:base64"]
+sources-netflow = ["sources-utils-net-udp", "tokio-util/net", "dep:base64", "dep:dashmap"]
 sources-nginx_metrics = ["dep:nom"]
 sources-opentelemetry = ["dep:hex", "vector-lib/opentelemetry", "dep:prost", "dep:prost-types", "sources-http_server", "sources-utils-http", "sources-utils-http-headers", "sources-vector"]
 sources-postgresql_metrics = ["dep:postgres-openssl", "dep:tokio-postgres"]
